Allow switching between float and double

+++ src/simple_mathlib.hpp
@@ -4,10 +4,11 @@
 #ifndef SIMPLE_MATHLIB_HPP
 #define SIMPLE_MATHLIB_HPP
 
-#include<stdint.h>
-#include<stdio.h>
-
+#include<cstdint>
+#include<cstdio>
 #include<cmath>
+
+#include "simple_texcomp.hpp"
 
 /* Endpoint interpolation constants */
 #define EP_LERP1  1.0 / 3.0
@@ -16,9 +17,9 @@
 /* Helper structs and math */
 struct Vec3f
 {
-    double x;
-    double y;
-    double z;
+    decimal x;
+    decimal y;
+    decimal z;
 
     inline Vec3f operator+(const Vec3f &other) const
     {
@@ -38,7 +39,7 @@ struct Vec3f
         };
     }
 
-    inline Vec3f operator-(double a) const
+    inline Vec3f operator-(decimal a) const
     {
         return Vec3f {
             x - a,
@@ -47,7 +48,7 @@ struct Vec3f
         };
     }
 
-    inline Vec3f operator*(double a) const
+    inline Vec3f operator*(decimal a) const
     {
         return Vec3f {
             x * a,
@@ -56,7 +57,7 @@ struct Vec3f
         };
     }
 
-    inline Vec3f operator/(double a) const
+    inline Vec3f operator/(decimal a) const
     {
         return Vec3f {
             x / a,
@@ -65,7 +66,7 @@ struct Vec3f
         };
     }
 
-    inline double dot(const Vec3f &other) const
+    inline decimal dot(const Vec3f &other) const
     {
         // return std::fmaf(x, other.x, std::fmaf(y, other.y, fmaf(z, other.z, 0.0)));
         return (x * other.x) + (y * other.y) + (z * other.z);
@@ -74,8 +75,8 @@ struct Vec3f
 
 struct Vec2f
 {
-    double x;
-    double y;
+    decimal x;
+    decimal y;
 
     inline Vec2f operator+(const Vec2f &other) const
     {
@@ -85,7 +86,7 @@ struct Vec2f
         };
     }
 
-    inline Vec2f operator+(double a) const
+    inline Vec2f operator+(decimal a) const
     {
         return Vec2f {
             x + a,
@@ -101,7 +102,7 @@ struct Vec2f
         };
     }
 
-    inline Vec2f operator-(double a) const
+    inline Vec2f operator-(decimal a) const
     {
         return Vec2f {
             x - a,
@@ -109,7 +110,7 @@ struct Vec2f
         };
     }
 
-    inline Vec2f operator*(double a) const
+    inline Vec2f operator*(decimal a) const
     {
         return Vec2f {
             x * a,
@@ -117,7 +118,7 @@ struct Vec2f
         };
     }
 
-    inline Vec2f operator/(double a) const
+    inline Vec2f operator/(decimal a) const
     {
         return Vec2f {
             x / a,
@@ -125,37 +126,37 @@ struct Vec2f
         };
     }
 
-    inline double dot(const Vec2f &other) const
+    inline decimal dot(const Vec2f &other) const
     {
         return (x * other.x) + (y * other.y);
     }
 };
 
-inline double fclamp(double a, double amin, double amax)
+inline decimal fclamp(decimal a, decimal amin, decimal amax)
 {
-    const double min = a < amin ? amin : a;
+    const decimal min = a < amin ? amin : a;
     return min > amax ? amax : min;
 }
 
 inline Vec3f min3f(const Vec3f &a, const Vec3f &b)
 {
     return Vec3f {
-        fmin(a.x, b.x),
-        fmin(a.y, b.y),
-        fmin(a.z, b.z),
+        (decimal)fmin(a.x, b.x),
+        (decimal)fmin(a.y, b.y),
+        (decimal)fmin(a.z, b.z),
     };
 }
 
 inline Vec3f max3f(const Vec3f &a, const Vec3f &b)
 {
     return Vec3f {
-        fmax(a.x, b.x),
-        fmax(a.y, b.y),
-        fmax(a.z, b.z),
+        (decimal)fmax(a.x, b.x),
+        (decimal)fmax(a.y, b.y),
+        (decimal)fmax(a.z, b.z),
     };
 }
 
-inline Vec3f clamp3f(const Vec3f &a, double amin, double amax)
+inline Vec3f clamp3f(const Vec3f &a, decimal amin, decimal amax)
 {
     return Vec3f {
         fclamp(a.x, amin, amax),
@@ -164,7 +165,7 @@ inline Vec3f clamp3f(const Vec3f &a, double amin, double amax)
     };
 }
 
-inline Vec2f clamp2f(const Vec2f &a, double amin, double amax)
+inline Vec2f clamp2f(const Vec2f &a, decimal amin, decimal amax)
 {
     return Vec2f {
         fclamp(a.x, amin, amax),
@@ -174,17 +175,17 @@ inline Vec2f clamp2f(const Vec2f &a, double amin, double amax)
 
 inline Vec2f abs2f(const Vec2f &a)
 {
-    return Vec2f { fabs(a.x), fabs(a.y) };
+    return Vec2f { (decimal)fabs(a.x), (decimal)fabs(a.y) };
 }
 
 /* Compute squared distance of two vectors */
-inline double distsq3f(const Vec3f &a, const Vec3f &b)
+inline decimal distsq3f(const Vec3f &a, const Vec3f &b)
 {
     Vec3f diff = a - b;
     return diff.dot(diff);
 }
 
-inline double distsq2f(const Vec2f &a, const Vec2f &b)
+inline decimal distsq2f(const Vec2f &a, const Vec2f &b)
 {
     Vec2f diff = a - b;
     return diff.dot(diff);
@@ -218,9 +219,9 @@ inline Vec3f rgb565_to_f32(uint16_t color)
     b = (b << 3) | (b >> 2);
 
     return Vec3f {
-        (double)(r) / 255.0,
-        (double)(g) / 255.0,
-        (double)(b) / 255.0,
+        (decimal)(r) / 255.0,
+        (decimal)(g) / 255.0,
+        (decimal)(b) / 255.0,
     };
 }
